+++ src/main/java/roomescape/admin/service/AdminReservationService.java
@@ -1,0 +1,22 @@
+package roomescape.admin.service;
+
+import java.util.List;
+import org.springframework.stereotype.Service;
+import roomescape.admin.controller.dto.AdminReservationResponse;
+import roomescape.reservation.repository.ReservationDao;
+
+@Service
+public class AdminReservationService {
+
+    private final ReservationDao reservationDao;
+
+    public AdminReservationService(ReservationDao reservationDao) {
+        this.reservationDao = reservationDao;
+    }
+
+    public List<AdminReservationResponse> getAllReservations() {
+        return reservationDao.findAll().stream()
+                .map(r -> AdminReservationResponse.from(r, r.getTheme()))
+                .toList();
+    }
+}
