+++ src/main/java/roomescape/admin/service/AdminThemeService.java
@@ -1,0 +1,23 @@
+package roomescape.admin.service;
+
+import org.springframework.stereotype.Service;
+import roomescape.theme.controller.dto.ThemeRequest;
+import roomescape.theme.repository.ThemeDao;
+
+@Service
+public class AdminThemeService {
+
+    private final ThemeDao themeDao;
+
+    public AdminThemeService(ThemeDao themeDao) {
+        this.themeDao = themeDao;
+    }
+
+    public Long save(ThemeRequest request) {
+        return themeDao.save(request.name(), request.description(), request.thumbnail_url());
+    }
+
+    public void delete(long id) {
+        themeDao.delete(id);
+    }
+}
