--- src/main/java/roomescape/domain/ReservationTime.java
@@ -1,24 +1,0 @@
-package roomescape.domain;
-
-import java.time.LocalTime;
-
-public class ReservationTime {
-    private final Long id;
-    private final LocalTime startAt;
-
-    public ReservationTime(Long id, LocalTime startAt) {
-        if (startAt == null) {
-            throw new IllegalArgumentException("예약 시간은 필수입니다.");
-        }
-        this.id = id;
-        this.startAt = startAt;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public LocalTime getStartAt() {
-        return startAt;
-    }
-}
