--- src/main/java/roomescape/dto/ReservationRequest.java
@@ -1,6 +1,0 @@
-package roomescape.dto;
-
-import java.time.LocalDate;
-
-public record ReservationRequest(String name, LocalDate date, Long timeId, Long themeId) {
-}
