--- src/main/java/roomescape/dto/ReservationResponse.java
@@ -1,22 +1,0 @@
-package roomescape.dto;
-
-import roomescape.domain.Reservation;
-import roomescape.domain.Theme;
-
-public record ReservationResponse(Long id,
-                                  String date,
-                                  String themeName,
-                                  String themeDescription,
-                                  String themeThumbnailUrl,
-                                  String time) {
-    public static ReservationResponse from(Reservation reservation, Theme theme) {
-        return new ReservationResponse(
-                reservation.getId(),
-                reservation.getDate().toString(),
-                theme.getName(),
-                theme.getDescription(),
-                theme.getThumbnailUrl(),
-                reservation.getTime().getStartAt().toString()
-        );
-    }
-}
