--- src/main/java/roomescape/dto/ReservationTimeRequest.java
@@ -1,6 +1,0 @@
-package roomescape.dto;
-
-import java.time.LocalTime;
-
-public record ReservationTimeRequest(LocalTime startAt) {
-}
