--- src/main/java/roomescape/dto/ReservationTimeResponse.java
@@ -1,9 +1,0 @@
-package roomescape.dto;
-
-import roomescape.domain.ReservationTime;
-
-public record ReservationTimeResponse(Long id, String startAt) {
-    public static ReservationTimeResponse from(ReservationTime time) {
-        return new ReservationTimeResponse(time.getId(), time.getStartAt().toString());
-    }
-}
