--- src/main/java/roomescape/dto/SearchRequest.java
@@ -1,7 +1,0 @@
-package roomescape.dto;
-
-public record SearchRequest(
-        String condition,
-        int size
-) {
-}
