--- src/main/java/roomescape/dto/ThemeRequest.java
@@ -1,4 +1,0 @@
-package roomescape.dto;
-
-public record ThemeRequest (String name, String description, String thumbnail_url){
-}
