--- src/main/java/roomescape/repository/ReservationTimeDao.java
@@ -1,54 +1,0 @@
-package roomescape.repository;
-
-import java.time.LocalTime;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Repository;
-import roomescape.domain.ReservationTime;
-
-@Repository
-public class ReservationTimeDao {
-    private final JdbcTemplate jdbcTemplate;
-    private final SimpleJdbcInsert jdbcInsert;
-
-    private final RowMapper<ReservationTime> timeRowMapper = (rs, rowNum) -> new ReservationTime(
-            rs.getLong("id"),
-            rs.getTime("start_at").toLocalTime()
-    );
-
-    public ReservationTimeDao(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-        this.jdbcInsert = new SimpleJdbcInsert(jdbcTemplate)
-                .withTableName("reservation_time")
-                .usingGeneratedKeyColumns("id");
-    }
-
-    public List<ReservationTime> findAll() {
-        return jdbcTemplate.query("SELECT id, start_at FROM reservation_time", timeRowMapper);
-    }
-
-    public Optional<ReservationTime> findById(Long id) {
-        List<ReservationTime> results = jdbcTemplate.query(
-                "SELECT id, start_at FROM reservation_time WHERE id = ?", timeRowMapper, id);
-        return results.stream().findFirst();
-    }
-
-    public boolean existsByStartAt(LocalTime startAt) {
-        Integer count = jdbcTemplate.queryForObject(
-                "SELECT COUNT(*) FROM reservation_time WHERE start_at = ?", Integer.class, startAt);
-        return count != null && count > 0;
-    }
-
-    public ReservationTime save(LocalTime startAt) {
-        Long id = jdbcInsert.executeAndReturnKey(Map.of("start_at", startAt)).longValue();
-        return new ReservationTime(id, startAt);
-    }
-
-    public void delete(Long id) {
-        jdbcTemplate.update("DELETE FROM reservation_time WHERE id = ?", id);
-    }
-}
