+++ src/main/java/roomescape/reservation/controller/ReservationController.java
@@ -1,0 +1,44 @@
+package roomescape.reservation.controller;
+
+import java.util.List;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.reservation.controller.dto.ReservationResponse;
+import roomescape.reservation.service.ReservationService;
+import roomescape.reservation.controller.dto.ReservationRequest;
+
+@RequestMapping("/reservations")
+@RestController
+public class ReservationController {
+    private final ReservationService reservationService;
+
+    public ReservationController(ReservationService reservationService) {
+        this.reservationService = reservationService;
+    }
+
+    @GetMapping
+    public ResponseEntity<List<ReservationResponse>> getReservations(@RequestParam String username) {
+        return ResponseEntity.ok(reservationService.findAllByName(username));
+    }
+
+    @ResponseStatus(HttpStatus.CREATED)
+    @PostMapping
+    public ReservationResponse createReservation(@RequestBody ReservationRequest request) {
+        return reservationService.save(request);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
+        reservationService.delete(id);
+        return ResponseEntity.noContent().build();
+    }
+}
