+++ src/main/java/roomescape/reservation/controller/dto/ReservationRequest.java
@@ -1,0 +1,6 @@
+package roomescape.reservation.controller.dto;
+
+import java.time.LocalDate;
+
+public record ReservationRequest(String name, LocalDate date, Long timeId, Long themeId) {
+}
