+++ src/main/java/roomescape/reservation/controller/dto/ReservationTimeRequest.java
@@ -1,0 +1,6 @@
+package roomescape.reservation.controller.dto;
+
+import java.time.LocalTime;
+
+public record ReservationTimeRequest(LocalTime startAt) {
+}
