+++ src/main/java/roomescape/reservation/controller/dto/ReservationTimeResponse.java
@@ -1,0 +1,9 @@
+package roomescape.reservation.controller.dto;
+
+import roomescape.reservation.domain.ReservationTime;
+
+public record ReservationTimeResponse(Long id, String startAt) {
+    public static ReservationTimeResponse from(ReservationTime time) {
+        return new ReservationTimeResponse(time.getId(), time.getStartAt().toString());
+    }
+}
