+++ src/main/java/roomescape/reservation/domain/Reservation.java
@@ -1,0 +1,53 @@
+package roomescape.reservation.domain;
+
+import java.time.LocalDate;
+import roomescape.theme.doamin.Theme;
+
+public class Reservation {
+    private final Long id;
+    private final String name;
+    private final LocalDate date;
+    private final ReservationTime time;
+    private final Theme theme;
+
+    public Reservation(Long id, String name, LocalDate date, ReservationTime time, Theme theme) {
+        validate(name, date, time);
+        this.id = id;
+        this.name = name;
+        this.date = date;
+        this.time = time;
+        this.theme = theme;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public LocalDate getDate() {
+        return date;
+    }
+
+    public ReservationTime getTime() {
+        return time;
+    }
+
+    public Theme getTheme() {
+        return theme;
+    }
+
+    public static void validate(String name, LocalDate date, ReservationTime time) {
+        if (name == null || name.isBlank()) {
+            throw new IllegalArgumentException("예약자 이름은 비어있을 수 없습니다.");
+        }
+        if (date == null) {
+            throw new IllegalArgumentException("예약 날짜는 필수입니다.");
+        }
+        if (time == null) {
+            throw new IllegalArgumentException("예약 시간은 필수입니다.");
+        }
+    }
+}
