+++ src/main/java/roomescape/reservation/service/ReservationTimeService.java
@@ -1,0 +1,41 @@
+package roomescape.reservation.service;
+
+import java.util.List;
+import org.springframework.stereotype.Service;
+import roomescape.reservation.domain.ReservationTime;
+import roomescape.reservation.repository.ReservationDao;
+import roomescape.reservation.repository.ReservationTimeDao;
+import roomescape.reservation.controller.dto.ReservationTimeRequest;
+import roomescape.reservation.controller.dto.ReservationTimeResponse;
+
+@Service
+public class ReservationTimeService {
+    private final ReservationTimeDao reservationTimeDao;
+    private final ReservationDao reservationDao;
+
+    public ReservationTimeService(ReservationTimeDao reservationTimeDao, ReservationDao reservationDao) {
+        this.reservationTimeDao = reservationTimeDao;
+        this.reservationDao = reservationDao;
+    }
+
+    public List<ReservationTimeResponse> findAll() {
+        return reservationTimeDao.findAll().stream()
+                .map(ReservationTimeResponse::from)
+                .toList();
+    }
+
+    public ReservationTimeResponse save(ReservationTimeRequest request) {
+        if (reservationTimeDao.existsByStartAt(request.startAt())) {
+            throw new IllegalArgumentException("이미 존재하는 예약 시간입니다.");
+        }
+        ReservationTime saved = reservationTimeDao.save(request.startAt());
+        return ReservationTimeResponse.from(saved);
+    }
+
+    public void delete(Long id) {
+        if (reservationDao.existsByTimeId(id)) {
+            throw new IllegalArgumentException("예약에 사용 중인 시간은 삭제할 수 없습니다.");
+        }
+        reservationTimeDao.delete(id);
+    }
+}
