--- src/main/java/roomescape/service/ReservationTimeService.java
@@ -1,41 +1,0 @@
-package roomescape.service;
-
-import java.util.List;
-import org.springframework.stereotype.Service;
-import roomescape.domain.ReservationTime;
-import roomescape.repository.ReservationDao;
-import roomescape.repository.ReservationTimeDao;
-import roomescape.dto.ReservationTimeRequest;
-import roomescape.dto.ReservationTimeResponse;
-
-@Service
-public class ReservationTimeService {
-    private final ReservationTimeDao reservationTimeDao;
-    private final ReservationDao reservationDao;
-
-    public ReservationTimeService(ReservationTimeDao reservationTimeDao, ReservationDao reservationDao) {
-        this.reservationTimeDao = reservationTimeDao;
-        this.reservationDao = reservationDao;
-    }
-
-    public List<ReservationTimeResponse> findAll() {
-        return reservationTimeDao.findAll().stream()
-                .map(ReservationTimeResponse::from)
-                .toList();
-    }
-
-    public ReservationTimeResponse save(ReservationTimeRequest request) {
-        if (reservationTimeDao.existsByStartAt(request.startAt())) {
-            throw new IllegalArgumentException("이미 존재하는 예약 시간입니다.");
-        }
-        ReservationTime saved = reservationTimeDao.save(request.startAt());
-        return ReservationTimeResponse.from(saved);
-    }
-
-    public void delete(Long id) {
-        if (reservationDao.existsByTimeId(id)) {
-            throw new IllegalArgumentException("예약에 사용 중인 시간은 삭제할 수 없습니다.");
-        }
-        reservationTimeDao.delete(id);
-    }
-}
