+++ src/main/java/roomescape/theme/controller/dto/ThemeRequest.java
@@ -1,0 +1,4 @@
+package roomescape.theme.controller.dto;
+
+public record ThemeRequest (String name, String description, String thumbnail_url){
+}
