+++ src/test/java/roomescape/reservationTime/ReservationTimeTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalTime;
 import org.junit.jupiter.api.Test;
-import roomescape.domain.ReservationTime;
+import roomescape.reservation.domain.ReservationTime;
 
 class ReservationTimeTest {
